IMPALA-6490: Reconnect shell when remote restarts
If the remote impalad died while the shell waited for a command to
complete, the shell disconnected. After restarting the remote impalad
(e.g. in a dev environment), it was necessary to reconnect the shell by
running "connect;".

This change makes the shell reconnect automatically in the precmd()
function.

I tested this manually and ran the existing shell tests to make sure
that the refactoring of _connect() does not introduce a regression.

Change-Id: I625c363aab44876f405c441523059b9cb8e0b9eb